+++ daemon-exp/ghc-specter-daemon-exp.cabal
@@ -20,7 +20,8 @@ source-repository head
 
 executable ghc-specter-daemon-exp
   main-is:       Main.hs
-  other-modules: ModuleGraph
+  other-modules: Handler
+                 ModuleGraph
                  Renderer
                  Util
                  Timing
